Device components reorg
- Reorganized components for devices lists, so that there are less files
