+++ packages/ui/src/components/lists/devices/EmailList/EmailList.tsx
@@ -1,19 +1,92 @@
-/**
- * Displays a list of emails for the user to manage
- */
 import { DeleteOneInput, DeleteType, Email, EmailCreateInput, emailValidation, endpointPostDeleteOne, endpointPostEmail, endpointPostEmailVerification, SendVerificationEmailInput, Success } from "@local/shared";
-import { IconButton, InputAdornment, Stack, useTheme } from "@mui/material";
+import { Box, IconButton, InputAdornment, ListItem, ListItemText, Stack, Tooltip, useTheme } from "@mui/material";
 import { fetchLazyWrapper } from "api";
 import { ListContainer } from "components/containers/ListContainer/ListContainer";
 import { TextInput } from "components/inputs/TextInput/TextInput";
 import { useFormik } from "formik";
 import { useLazyFetch } from "hooks/useLazyFetch";
-import { AddIcon, EmailIcon } from "icons";
+import { AddIcon, CompleteIcon, DeleteIcon, EmailIcon } from "icons";
 import { useCallback } from "react";
 import { useTranslation } from "react-i18next";
+import { multiLineEllipsis } from "styles";
 import { PubSub } from "utils/pubsub";
-import { EmailListItem } from "../EmailListItem/EmailListItem";
-import { EmailListProps } from "../types";
+import { EmailListItemProps, EmailListProps } from "../types";
+
+const Status = {
+    NotVerified: "#a71c2d", // Red
+    Verified: "#19972b", // Green
+};
+
+/**
+ * Displays a list of emails for the user to manage
+ */
+export function EmailListItem({
+    handleDelete,
+    handleVerify,
+    index,
+    data,
+}: EmailListItemProps) {
+    const { palette } = useTheme();
+
+    const onDelete = useCallback(() => {
+        handleDelete(data);
+    }, [data, handleDelete]);
+
+    const onVerify = useCallback(() => {
+        handleVerify(data);
+    }, [data, handleVerify]);
+
+    return (
+        <ListItem
+            disablePadding
+            sx={{
+                display: "flex",
+                padding: 1,
+                borderBottom: `1px solid ${palette.divider}`,
+            }}
+        >
+            {/* Left informational column */}
+            <Stack direction="column" spacing={1} pl={2} sx={{ marginRight: "auto" }}>
+                <ListItemText
+                    primary={data.emailAddress}
+                    sx={{ ...multiLineEllipsis(1) }}
+                />
+                {/* Verified indicator */}
+                <Box sx={{
+                    borderRadius: 1,
+                    border: `2px solid ${data.verified ? Status.Verified : Status.NotVerified}`,
+                    color: data.verified ? Status.Verified : Status.NotVerified,
+                    height: "fit-content",
+                    fontWeight: "bold",
+                    marginTop: "auto",
+                    marginBottom: "auto",
+                    textAlign: "center",
+                    padding: 0.25,
+                    width: "fit-content",
+                }}>
+                    {data.verified ? "Verified" : "Not Verified"}
+                </Box>
+            </Stack>
+            {/* Right action buttons */}
+            <Stack direction="row" spacing={1}>
+                {!data.verified && <Tooltip title="Resend email verification">
+                    <IconButton
+                        onClick={onVerify}
+                    >
+                        <CompleteIcon fill={Status.NotVerified} />
+                    </IconButton>
+                </Tooltip>}
+                <Tooltip title="Delete Email">
+                    <IconButton
+                        onClick={onDelete}
+                    >
+                        <DeleteIcon fill={palette.secondary.main} />
+                    </IconButton>
+                </Tooltip>
+            </Stack>
+        </ListItem>
+    );
+}
 
 export const EmailList = ({
     handleUpdate,
