--- packages/ui/src/components/lists/devices/email/index.ts
@@ -1,3 +1,0 @@
-export * from "./EmailList/EmailList";
-export * from "./EmailListItem/EmailListItem";
-
