--- packages/ui/src/components/lists/devices/index.ts
@@ -1,4 +1,0 @@
-export * from "./email";
-export * from "./push";
-export * from "./wallet";
-
