--- packages/ui/src/components/lists/devices/push/index.ts
@@ -1,3 +1,0 @@
-export * from "./PushList/PushList";
-export * from "./PushListItem/PushListItem";
-
