--- packages/ui/src/components/lists/devices/wallet/index.ts
@@ -1,3 +1,0 @@
-export * from "./WalletList/WalletList";
-export * from "./WalletListItem/WalletListItem";
-
